+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog for RapyutaSimulationPlugins repository
 
 ## 0.0.9 ##
-* `SimulationState` Add `SpawnEntity()`, used by `SpawnEntitySrv + SpawnEntityListSrv`
+* `SimulationState` Add `SpawnEntity()`, used by `SpawnEntitySrv + SpawnEntitiesSrv`
 
 ## 0.0.8 ##
 * Add ARRBaseRobot, ARRRobotROS2Interface taking src from ARobotROSController